Stippled stripes
Stipples look not so great in high density. The method of creation
becomes apparent. It would be better if I actually computed their
locations better --- the right way is definitely to place points while
checking that each point is no closer than delta to the nearest point.
The problem is that this is naively very slow.

A faster way to do this would be to partition each rectangle into grids,
and check only those points within those grid partitions adjacent to the
proposed point. This seems too anoying right now --- so this is
something for a future challenge.